Move read single event endpoint to routes directory
